% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractsClusterMakeR.R
\name{AbstractsClusterMakeR}
\alias{AbstractsClusterMakeR}
\title{Make Preliminary Network for QC}
\usage{
AbstractsClusterMakeR(AbstractStrings, IDs)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}

\item{IDs}{character or numeric vector: unique abstract IDs}
}
\description{
Make Preliminary Network for QC
}
