% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractsGetteR.R
\name{AbstractsGetteR}
\alias{AbstractsGetteR}
\title{Gets Abstracts Frm PubMed IDs}
\usage{
AbstractsGetteR(IDs)
}
\arguments{
\item{IDs}{character or numeric vector: unique abstract IDs}
}
\description{
Gets Abstracts Frm PubMed IDs
}
