% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractsProportionCalculator.R
\name{AbstractsProportionCalculator}
\alias{AbstractsProportionCalculator}
\title{Calculates Proportion of Abstracts from which Results Were Obtained}
\usage{
AbstractsProportionCalculator(Result, IDs)
}
\arguments{
\item{Result}{data.frame: either genes, phenotypes, or species results}

\item{IDs}{character or numeric vector: unique abstract IDs}
}
\description{
Calculates Proportion of Abstracts from which Results Were Obtained
}
