% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaNumericalizeR.R
\name{AlphaNumericalizeR}
\alias{AlphaNumericalizeR}
\title{Removes Non-Alphanumerical Characters From Abstracts}
\usage{
AlphaNumericalizeR(AbstractStrings)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}
}
\description{
Removes Non-Alphanumerical Characters From Abstracts
}
