% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnimalGeneSynonymyKey}
\alias{AnimalGeneSynonymyKey}
\title{Key Associating Non-Overlapping Synonyms to Accepted Animals Gene Names}
\format{
A data frame with 36661 rows and 4 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
AnimalGeneSynonymyKey
}
\description{
Key Associating Non-Overlapping Synonyms to Accepted Animals Gene Names
}
\keyword{datasets}
