% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnimalSpp}
\alias{AnimalSpp}
\title{Animal Species Library}
\format{
A data frame with 1060491 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
AnimalSpp
}
\description{
Animal Species Library
}
\keyword{datasets}
