% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnimalSwissGenesCombo}
\alias{AnimalSwissGenesCombo}
\title{SwissProt Gene Library for Animals}
\format{
A data frame with 74610 rows and 3 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
AnimalSwissGenesCombo
}
\description{
SwissProt Gene Library for Animals
}
\keyword{datasets}
