% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnimalTerms}
\alias{AnimalTerms}
\title{Curated and Expanded Glossary of Animal Terms}
\format{
A data frame with 257 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://animalcorner.org/glossary/}
}
\usage{
AnimalTerms
}
\description{
Curated and Expanded Glossary of Animal Terms
}
\keyword{datasets}
