% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AroundWordsSearcheR.R
\name{AroundWordsSearcheR}
\alias{AroundWordsSearcheR}
\title{N-gram abstract word search function}
\usage{
AroundWordsSearcheR(AbstractStrings, IDs, Anchors, n, RemoveCommons)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{Anchors}{haracter vector: a vector of words for which to search}

\item{n}{numeric: the number of words before and after anchors to search}

\item{RemoveCommons}{boolean: remove common English words}
}
\description{
N-gram abstract word search function
}
