% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanPhenotypesLibrary.R
\name{CleanPhenotypesLibrary}
\alias{CleanPhenotypesLibrary}
\title{Cleans Phenotypes Library}
\usage{
CleanPhenotypesLibrary(MoBotTerms)
}
\arguments{
\item{MoBotTerms}{character vector: all phenotypes as strings}
}
\description{
Cleans Phenotypes Library
}
