% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonWordsRemoveR.R
\name{CommonWordsRemoveR}
\alias{CommonWordsRemoveR}
\title{Checks If Gene Matches Are English Words}
\usage{
CommonWordsRemoveR(String)
}
\arguments{
\item{String}{character: a string}
}
\description{
Checks If Gene Matches Are English Words
}
