% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusInferreR.R
\name{ConsensusInferreR}
\alias{ConsensusInferreR}
\title{Explore Inter-Dataset Intersectionality}
\usage{
ConsensusInferreR(Ta, G, P, AbstractsSpp, GenesOut, AbsPhen)
}
\arguments{
\item{Ta}{data.frame: longform taxonomical results OR NULL}

\item{G}{data.frame: longform genes results OR NULL}

\item{P}{data.frame: longform phenotypes results OR NULL}

\item{AbstractsSpp}{data.frame: shortform taxonomical results OR NULL}

\item{GenesOut}{data.frame: shortform genes results OR NULL}

\item{AbsPhen}{data.frame: shortform phenotypes results OR NULL}
}
\description{
Explore Inter-Dataset Intersectionality
}
