% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOccBringeR.R
\name{CoOcBringeR}
\alias{CoOcBringeR}
\title{Gets Abstracts IDs From Results Data Using Search Term}
\usage{
CoOcBringeR(Terms1, Terms2, SearchingData1, SearchingData2, IDs, Type)
}
\arguments{
\item{Terms1}{character or numeric vector: unique abstract IDs}

\item{Terms2}{character or numeric vector: unique abstract IDs}

\item{SearchingData1}{character or numeric vector: unique abstract IDs}

\item{SearchingData2}{character or numeric vector: unique abstract IDs}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{Type}{charcater: either S, G, or P (Species, Genes, or Species Analysis)}
}
\description{
Gets Abstracts IDs From Results Data Using Search Term
}
