% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateTaxonomyGBIF.R
\name{CurateTaxonomy}
\alias{CurateTaxonomy}
\title{Curates Taxonomy By Checking For Synonymy And Publication Validity}
\usage{
CurateTaxonomy(list_of_spp)
}
\arguments{
\item{mainDirect}{Main directory}

\item{spp}{a data.frame of species "Genus spp"}

\item{nomen}{chosen file name for output}
}
\description{
Curates Taxonomy By Checking For Synonymy And Publication Validity
}
