% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomLookeR.R
\name{CustomLookeR}
\alias{CustomLookeR}
\title{Extracts Custom Words From Abstracts}
\usage{
CustomLookeR(AbstractStrings, IDs, CustomWords, CustomCategories)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{CustomWords}{character vector: user terms to use for search library}

\item{CustomCategories}{character vector: user categories to use for search library}
}
\description{
Extracts Custom Words From Abstracts
}
