% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyG2PMineR.R
\name{EasyG2PMineR}
\alias{EasyG2PMineR}
\title{Easy G2PMineR Analysis Function}
\usage{
EasyG2PMineR(Query, Kingdom, Seed, HowMany)
}
\arguments{
\item{Query}{character: Search terms}

\item{Kingdom}{character: which kingdom to use}

\item{Seed}{numeric: a number}

\item{HowMany}{numeric: max number of abstracts to sample}
}
\description{
Easy G2PMineR Analysis Function
}
