% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FungiGeneSynonymyKey}
\alias{FungiGeneSynonymyKey}
\title{Key Associating Non-Overlapping Synonyms to Accepted Fungi Gene Names}
\format{
A data frame with 4442 rows and 4 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
FungiGeneSynonymyKey
}
\description{
Key Associating Non-Overlapping Synonyms to Fungi Animals Gene Names
}
\keyword{datasets}
