% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FungiSpp}
\alias{FungiSpp}
\title{Fungi Species Library}
\format{
A data frame with 120578 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
FungiSpp
}
\description{
Fungi Species Library
}
\keyword{datasets}
