% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FungiSwissGenesCombo}
\alias{FungiSwissGenesCombo}
\title{SwissProt Gene Library for Fungi}
\format{
A data frame with 14136 rows and 3 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
FungiSwissGenesCombo
}
\description{
SwissProt Gene Library for Fungi
}
\keyword{datasets}
