% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FungiTerms}
\alias{FungiTerms}
\title{Curated and Expanded Glossary of Fungi Terms}
\format{
A data frame with 163 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://animalcorner.org/glossary/}
}
\usage{
FungiTerms
}
\description{
Curated and Expanded Glossary of Fungi Terms
}
\keyword{datasets}
