% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneFamiliesAssociatoR.R
\name{GeneFamiliesAssociatoR}
\alias{GeneFamiliesAssociatoR}
\title{Preprocesses genes (OPTIONAL)}
\usage{
GeneFamiliesAssociatoR(SwissGenesRaw)
}
\arguments{
\item{SwissGenesRaw}{matrix: swissprot genes}
}
\description{
Preprocesses genes (OPTIONAL)
}
