% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneFrequencySifteR.R
\name{GeneFrequencySifteR}
\alias{GeneFrequencySifteR}
\title{Siftes Genes by Gene Frequency}
\usage{
GeneFrequencySifteR(GenesOut, IDs)
}
\arguments{
\item{GenesOut}{matrix: swissprot genes abstracts search results}

\item{IDs}{vector: unique identifiers}
}
\description{
Siftes Genes by Gene Frequency
}
