% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNamesGroupeR.R
\name{GeneNamesGroupeR}
\alias{GeneNamesGroupeR}
\title{Creates Artificial Gene Broups Based On Non-Numerical Names}
\usage{
GeneNamesGroupeR(input)
}
\arguments{
\item{input}{character vector or data.frame: input}
}
\description{
Creates Artificial Gene Broups Based On Non-Numerical Names
}
