% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesLookeR.R
\name{GenesLookeR}
\alias{GenesLookeR}
\title{Mines Abstract for Gene Names}
\usage{
GenesLookeR(AbstractStrings, IDs, Kingdom, Add, SppAbbr)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{Kingdom}{character: which kingdom to use}

\item{Add}{data.frame: user gene names, families, and ontologies to add to library}

\item{SppAbbr}{character vector: list of species abbreviations to use to make isoforms}
}
\description{
Mines Abstract for Gene Names
}
