% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneSynonymyKey}
\alias{GeneSynonymyKey}
\title{Key Associating Non-Overlapping Synonyms to Accepted Gene Names}
\format{
A data frame with 7288 rows and 4 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
GeneSynonymyKey
}
\description{
Key Associating Non-Overlapping Synonyms to Accepted Gene Names
}
\keyword{datasets}
