% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLElementRemoveR.R
\name{HTMLElementRemoveR}
\alias{HTMLElementRemoveR}
\title{RemovesHTML elements}
\usage{
HTMLElementRemoveR(AbstractsStrings)
}
\arguments{
\item{AbstractsStrings}{character vector: character vector: of abstracts as strings}
}
\description{
RemovesHTML elements
}
