% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalPairwiseDistanceInferreR.R
\name{InternalPairwiseDistanceInferreR}
\alias{InternalPairwiseDistanceInferreR}
\title{Turn Distance Matrix into Longform Table}
\usage{
InternalPairwiseDistanceInferreR(Things, Matches, allabsnum)
}
\arguments{
\item{Things}{character vector: strings}

\item{Matches}{character vector: matches, comma separated}

\item{allabsnum}{numeric: total number of abstracts in analysis}
}
\description{
Turn Distance Matrix into Longform Table
}
