% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookeRBringeR.R
\name{LookeRBringeR}
\alias{LookeRBringeR}
\title{Gets Abstracts IDs From Results Data Using Search}
\usage{
LookeRBringeR(Terms, SearchingData, IDs, Type)
}
\arguments{
\item{Terms}{character or numeric vector: unique abstract IDs}

\item{SearchingData}{character or numeric vector: unique abstract IDs}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{Type}{charcater: either S, G, or P (Species, Genes, or Species Analysis)}
}
\description{
Gets Abstracts IDs From Results Data Using Search
}
