% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAbsPhenLongform.R
\name{MakeAbsPhenLongform}
\alias{MakeAbsPhenLongform}
\title{Makes AbstractsSpp Longform}
\usage{
MakeAbsPhenLongform(AbsPhen)
}
\arguments{
\item{AbsPhen}{data.frame: output of SpeciesLookeR}
}
\description{
Makes AbstractsSpp Longform
}
