% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAbstractsSppLongform.R
\name{MakeAbstractsSppLongform}
\alias{MakeAbstractsSppLongform}
\title{Makes AbstractsSpp Longform}
\usage{
MakeAbstractsSppLongform(AbstractsSpp)
}
\arguments{
\item{AbstractsSpp}{data.frame: output of SpeciesLookeR}
}
\description{
Makes AbstractsSpp Longform
}
