% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGenesOutLongform.R
\name{MakeGenesOutLongform}
\alias{MakeGenesOutLongform}
\title{Makes GenesOut Longform}
\usage{
MakeGenesOutLongform(GenesOut)
}
\arguments{
\item{GenesOut}{data.frame: output of SynonymReplaceR}
}
\description{
Makes GenesOut Longform
}
