% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchesBarPlotteR.R
\name{MatchesBarPlotteR}
\alias{MatchesBarPlotteR}
\title{Makes Matches Barplot}
\usage{
MatchesBarPlotteR(strings, matches, n)
}
\arguments{
\item{strings}{character vector: abstracts text}

\item{matches}{character vector: comma separated matched IDs}

\item{n}{numeric: how many columns, either numeric or character of "all"}
}
\description{
Makes Matches Barplot
}
