% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipInvestigatoR.R
\name{MembershipInvestigatoR}
\alias{MembershipInvestigatoR}
\title{Checks If Gene Matches Are English Words}
\usage{
MembershipInvestigatoR(membership, threshold, singularize)
}
\arguments{
\item{membership}{data.frame or matrix: output of AbstractsClusterMakeR}

\item{threshold}{numerical: between 0 and 1, proportion of abstracts that must contain match to be considered widespread}

\item{singularize}{boolean: if T then all words are singularized}
}
\description{
Checks If Gene Matches Are English Words
}
