% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MoBotTerms}
\alias{MoBotTerms}
\title{Curated and Expanded Mobot Glossary of Plant Terms}
\format{
A data frame with 2986 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{http://www.mobot.org/MOBOT/Research/APweb/top/glossarya_h.html/}
}
\usage{
MoBotTerms
}
\description{
Curated and Expanded Mobot Glossary of Plant Terms
}
\keyword{datasets}
