% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseDisctanceInferreR.R
\name{PairwiseDistanceInferreR}
\alias{PairwiseDistanceInferreR}
\title{Finds weighted pairwise distances between two levels using shared abstracts}
\usage{
PairwiseDistanceInferreR(terms1, matches1, terms2, matches2, allabsnum)
}
\arguments{
\item{terms1}{a vector of unique strings}

\item{matches1}{a vector of unique identifiers for abstracts collated by commas for each term}

\item{terms2}{a vector of unique strings}

\item{matches2}{a vector of unique identifiers for abstracts collated by commas for each term}

\item{allabsnum}{total number of abstracts in analysis}
}
\description{
Finds weighted pairwise distances between two levels using shared abstracts
}
