% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeLookeR.R
\name{PhenotypeLookeR}
\alias{PhenotypeLookeR}
\title{Extracts Phenotypes From Abstracts}
\usage{
PhenotypeLookeR(AbstractStrings, IDs, Kingdom, Add)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{Kingdom}{character: which kingdom to use}

\item{Add}{character vector: user terms to add to phenology library}
}
\description{
Extracts Phenotypes From Abstracts
}
