% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesAbbreviatoR.R
\name{SpeciesAbbreviatoR}
\alias{SpeciesAbbreviatoR}
\title{Gets Abstracts IDs From Results Data Using Search}
\usage{
SpeciesAbbreviatoR(AbstractsSpp)
}
\arguments{
\item{AbstractsSpp}{data.frame: output of SpeciesLookeR}
}
\description{
Gets Abstracts IDs From Results Data Using Search
}
