% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesLookeR.R
\name{SpeciesLookeR}
\alias{SpeciesLookeR}
\title{Checks If Species is in String}
\usage{
SpeciesLookeR(AbstractStrings, IDs, Kingdom, Add)
}
\arguments{
\item{AbstractStrings}{character vector: of abstracts as strings}

\item{IDs}{character or numeric vector: unique abstract IDs}

\item{Kingdom}{character: which kingdom to use}

\item{Add}{character vector: user terms to add to phenology library}
}
\description{
Checks If Species is in String
}
