% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesPrepareR.R
\name{SpeciesPrepareR}
\alias{SpeciesPrepareR}
\title{Checks If Species is in String}
\usage{
SpeciesPrepareR(AbstractSpecies)
}
\arguments{
\item{AbstractSpecies}{a vector of strings}
}
\description{
Checks If Species is in String
}
