% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SwissGenesCombo}
\alias{SwissGenesCombo}
\title{SwissProt Gene Library}
\format{
A data frame with 21733 rows and 3 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
SwissGenesCombo
}
\description{
SwissProt Gene Library
}
\keyword{datasets}
