% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynonymReplaceR.R
\name{SynonymReplaceR}
\alias{SynonymReplaceR}
\title{Replaces Gene Synonyms With Accepted Nomenclature}
\usage{
SynonymReplaceR(GenesOut, Kingdom)
}
\arguments{
\item{GenesOut}{matrix: swissprot genes abstracts search results}

\item{Kingdom}{boolean: which kingdom to use}
}
\description{
Replaces Gene Synonyms With Accepted Nomenclature
}
