% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxoGenoPhenoIntegratoR.R
\name{TaxoGenoPhenoIntegratoR}
\alias{TaxoGenoPhenoIntegratoR}
\title{Integrate Matches For Taxonomy, Genes, and Phenotypes Analyses}
\usage{
TaxoGenoPhenoIntegratoR(AbstractsSpp, GenesOut, AbsPhen, MakeVenn)
}
\arguments{
\item{AbstractsSpp}{data.frame: output of SpeciesLookeR}

\item{GenesOut}{data.frame: output of GenesLookeR}

\item{AbsPhen}{data.frame: output of PhenotypeLookeR}

\item{MakeVenn}{boolean: make venn diagram of intersection?}
}
\description{
Integrate Matches For Taxonomy, Genes, and Phenotypes Analyses
}
