% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopN_PickeR.R
\name{TopN_PickeR}
\alias{TopN_PickeR}
\title{Picks n Most Similar Pairs}
\usage{
TopN_PickeR(matrix, n, decreasing)
}
\arguments{
\item{matrix}{data.frame or matrix: a distance matrix dataframe}

\item{n}{numeric: how many tops t take a numeric}

\item{decreasing}{boolean: order of output}
}
\description{
Picks n Most Similar Pairs
}
