% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopN_Picker_Internal.R
\name{TopN_PickeR_Internal}
\alias{TopN_PickeR_Internal}
\title{Picks n Most Similar Pairs}
\usage{
TopN_PickeR_Internal(matrix, n, decreasing)
}
\arguments{
\item{matrix}{a distance matrix dataframe}

\item{n}{how many tops t take a numeric}

\item{decreasing}{boolean: order of output}
}
\description{
Picks n Most Similar Pairs
}
