% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TracheophytaSpp}
\alias{TracheophytaSpp}
\title{Trachephyta Species Libray}
\format{
A data frame with 300771 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
TracheophytaSpp
}
\description{
Trachephyta Species Libray
}
\keyword{datasets}
