% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityGradeR.R
\name{UtilityGradeR}
\alias{UtilityGradeR}
\title{Assigns A Grade To Gene Matches Based On Utility}
\usage{
UtilityGradeR(GenesOut, Kingdom, Add, Groups)
}
\arguments{
\item{GenesOut}{character or numeric vector: unique abstract IDs}

\item{Kingdom}{character: which kingdom to use}

\item{Add}{data.frame: user gene names, families, and ontologies to add to library}

\item{Groups}{data.frame: gene groups, one column output of GeneNamesGroupeR}
}
\description{
Assigns A Grade To Gene Matches Based On Utility
}
