% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VignetteAbstractsStrings}
\alias{VignetteAbstractsStrings}
\title{Abstracts Strings for Vignette}
\format{
A data frame with 1000 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
VignetteAbstractsStrings
}
\description{
Key Associating Non-Overlapping Synonyms to Fungi Animals Gene Names
}
\keyword{datasets}
