% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Webster}
\alias{Webster}
\title{Webster's Dictinary Words}
\format{
A data frame with 466550 rows and 1 variables:
\describe{
  \item{\code{2}}{}
}
}
\source{
\url{http://https://github.com/adambom/dictionary/}
}
\usage{
Webster
}
\description{
Webster's Dictinary Words
}
\keyword{datasets}
